Default Toto to float32 to satisfy Categorical Simplex constraint
Toto's output head samples from torch.distributions.Categorical over 24 bins.
In bfloat16 the per-bin probabilities don't sum to exactly 1.0, failing the
Simplex constraint check. Float32 has enough precision for the sum to round
correctly.

Authored-By: Rune Skardhamar <rune@predictioninstitute.com>

+++ src/gorm_ai/prediction/engines/toto_direct_engine.py
@@ -68,7 +68,11 @@ class TotoDirectEngine(ChronosPipelineEngine):
     _DEFAULT_PREFIX = "Datadog/"
 
     def __init__(self, model_id: str = DEFAULT_MODEL_ID):
-        super().__init__(model_id=model_id, default_precision="bfloat16")
+        # Toto's Categorical output head is precision-sensitive: bfloat16 sums
+        # of the per-bin probabilities don't exactly equal 1.0, which fails
+        # torch.distributions.Categorical's Simplex constraint.  Default to
+        # float32; users can still override via the "precision" DB parameter.
+        super().__init__(model_id=model_id, default_precision="float32")
         self._forecaster = None
         self._device: str = "cpu"
         self._torch_dtype = None  # set in _load_model()
